implementing replicator plugin

+++ The Interpolator/Assets/StormancerNetworkIdentity.cs
@@ -7,7 +7,7 @@ namespace Stormancer
     public class StormancerNetworkIdentity : MonoBehaviour
     {
         public bool IsMaster = false;
-        public long Id;
+        public uint Id;
         public int PrefabId;
 
         public List<SynchBehaviourBase> SynchBehaviours { get; set; }
